+++ lemon/extradmin/admin.py
@@ -46,6 +46,13 @@ class MenuSectionAdmin(extradmin.ModelAdmin):
 
 class UserExtrAdmin(extradmin.ModelAdmin, UserAdmin):
 
+    fieldsets = (
+        (None, {'fields': ('username',)}),
+        (_('Personal info'), {'fields': ('first_name', 'last_name', 'email')}),
+        (_('Permissions'), {'fields': ('is_active', 'is_staff', 'is_superuser',
+                                       'groups', 'user_permissions')}),
+        (_('Important dates'), {'fields': ('last_login', 'date_joined')}),
+    )
     add_fieldsets = (
         (None, {
             'classes': ('wide',),