+++ lemon/extradmin/migrations/0001_initial.py
@@ -1,13 +1,13 @@
-# encoding: utf-8
+# -*- coding: utf-8 -*-
 import datetime
 from south.db import db
 from south.v2 import SchemaMigration
 from django.db import models
 
+
 class Migration(SchemaMigration):
 
     def forwards(self, orm):
-        
         # Adding model 'MenuSection'
         db.create_table('extradmin_menusection', (
             ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
@@ -26,15 +26,12 @@ class Migration(SchemaMigration):
         ))
         db.send_create_signal('extradmin', ['MenuItem'])
 
-
     def backwards(self, orm):
-        
         # Deleting model 'MenuSection'
         db.delete_table('extradmin_menusection')
 
         # Deleting model 'MenuItem'
         db.delete_table('extradmin_menuitem')
-
 
     models = {
         'contenttypes.contenttype': {