+++ lemon/extradmin/options.py
@@ -2,6 +2,7 @@ import inspect
 
 from django import forms
 from django.contrib.admin import options
+from django.contrib.admin.templatetags.admin_static import static
 from django.contrib.admin.views.main import IS_POPUP_VAR
 from django.contrib.admin.widgets import AdminRadioSelect
 from django.conf import settings
@@ -98,90 +99,73 @@ class ModelAdmin(options.ModelAdmin, BaseModelAdmin):
     tabs = False
     action_description_overrides = {}
 
-    def __init__(self, model, admin_site):
-        super(ModelAdmin, self).__init__(model, admin_site)
-        if self.tabs:
-            if isinstance(self.tabs, bool):
-                tabs = []
-                for inline in self.inline_instances:
-                    tabs.append({'title': inline.verbose_name_plural,
-                                 'contents': [inline]})
-                self.tabs = tabs
-            else:
-                tabs = []
-                for tab in self.tabs:
-                    tabs.append(self._inspect_tab(tab))
-                self.tabs = tabs
-
-    def _inspect_tab(self, tab):
-        if inspect.isclass(tab) and issubclass(tab, options.InlineModelAdmin):
-            title, contents = self._get_tab_for_inline(tab)
-            return {'title': title, 'contents': contents}
+    def _inspect_tab(self, request, tab):
         if isinstance(tab, dict):
             contents = tab['contents']
-            if inspect.isclass(contents) and \
-               issubclass(contents, options.InlineModelAdmin):
-                title, contents = self._get_tab_for_inline(contents)
-                if 'title' in tab:
-                    title = tab['title']
-                return {'title': title, 'contents': contents}
-            if isinstance(contents, (list, tuple)):
-                if len(contents) == 1 and inspect.isclass(contents[0]) and \
-                   issubclass(contents[0], options.InlineModelAdmin):
-                    title, contents = self._get_tab_for_inline(contents[0])
-                    if 'title' in tab:
-                        title = tab['title']
-                    return {'title': title, 'contents': contents}
-                if 'title' not in tab:
-                    raise IncorrectTabFormat(
-                        u'No title provided for tab with multiple items.')
-                new_contents = []
-                for tab_item in contents:
-                    new_contents.extend(self._get_tab_for_inline(tab_item)[1])
-                tabs.append({'title': tab['title'], 'contents': new_contents})
-        raise IncorrectTabFormat(u'Incorrect tab format.')
+            if not isinstance(contents, (list, tuple)):
+                contents = [contents]
+            new_contents = []
+            for inline in contents:
+                inline_instance = self._get_inline_instance(request, inline)
+                new_contents.append(inline_instance)
+            return {'title': tab['title'], 'contents': new_contents}
+        title, contents = self._get_tab_for_inline(request, tab)
+        return {'title': title, 'contents': contents}
 
-    def _get_tab_for_inline(self, inline):
-        for inline_instance in self.inline_instances:
+    def _get_inline_instance(self, request, inline):
+        for inline_instance in self.get_inline_instances(request):
             if isinstance(inline_instance, inline):
-                title = inline_instance.verbose_name_plural
-                contents = [inline_instance]
-                break
-        else:
-            raise InlineInstanseNotFound(
-                u'It seems that inline %s is in %s.tabs but is not in '
-                u'%s.inlines.' % (inline.__name__, self.__class__.__name__,
-                                  self.__class__.__name__))
-        return title, contents
+                return inline_instance
+        raise InlineInstanseNotFound(u'It seems that inline %s is in %s.tabs '
+            u'but is not in %s.inlines.' % (inline.__name__,
+            self.__class__.__name__, self.__class__.__name__))
 
-    def _media(self):
-        js = [settings.ADMIN_MEDIA_PREFIX + 'js/SelectBox.js',
-              settings.STATIC_URL + 'extradmin/js/jquery.relatedobjectlookup.js']
+    def _get_tab_for_inline(self, request, inline):
+        inline_instance = self._get_inline_instance(request, inline)
+        return inline_instance.verbose_name_plural, [inline_instance]
+
+    @property
+    def media(self):
+        js = [static('admin/js/SelectBox.js'),
+              static('extradmin/js/jquery.relatedobjectlookup.js')]
         if self.prepopulated_fields:
-            js.extend([settings.ADMIN_MEDIA_PREFIX + 'js/urlify.js',
-                       settings.STATIC_URL + 'extradmin/js/jquery.prepopulate.js'])
+            js.extend(static('admin/js/urlify.js'),
+                      static('extradmin/js/jquery.prepopulate.js'))
         if self.tabs:
-            js.extend([settings.STATIC_URL + 'extradmin/js/jquery-ui.lemon.tabs.js'])
+            js.append(static('extradmin/js/jquery-ui.lemon.tabs.js'))
         return forms.Media(js=js)
-    media = property(_media)
+
+    def get_tabs(self, request):
+        if not self.tabs:
+            return False
+        tabs = []
+        if isinstance(self.tabs, bool):
+            for inline in self.get_inline_instances(request):
+                tabs.append({'title': inline.verbose_name_plural,
+                             'contents': [inline]})
+        else:
+            for tab in self.tabs:
+                tabs.append(self._inspect_tab(request, tab))
+        return tabs
 
     def render_change_form(self, request, context, add=False, change=False,
                            form_url='', obj=None):
-        tabs = False
+        tabs = self.get_tabs(request)
         tabulated_formsets = []
         inline_admin_formsets = context.get('inline_admin_formsets')
-        if self.tabs:
+        if tabs:
             formsets = {}
             for formset in context.get('inline_admin_formsets'):
-                formsets[formset.opts] = formset
-            tabs = []
-            for tab in self.tabs:
-                contents = [formsets[inline] for inline in tab['contents']]
-                tabs.append({'title': tab['title'], 'contents': contents})
+                formsets[formset.opts.__class__] = formset
+            new_tabs = []
+            for tab in tabs:
+                contents = [formsets[inline.__class__] for inline in tab['contents']]
+                new_tabs.append({'title': tab['title'], 'contents': contents})
                 tabulated_formsets.extend(contents)
-            contents = [formset for formset in inline_admin_formsets \
-                            if formset not in tabulated_formsets]
-            tabs[0:0] = [{'title': _('General'), 'contents': contents}]
+            contents = [formset for formset in inline_admin_formsets
+                        if formset not in tabulated_formsets]
+            new_tabs.insert(0, {'title': _('General'), 'contents': contents})
+            tabs = new_tabs
         context.update({'tabs': tabs})
         return super(ModelAdmin, self).render_change_form(
             request, context, add, change, form_url, obj)
@@ -251,13 +235,13 @@ class InlineModelAdmin(options.InlineModelAdmin, BaseModelAdmin):
     def __init__(self, parent_model, admin_site):
         super(InlineModelAdmin, self).__init__(parent_model, admin_site)
 
-    def _media(self):
-        js = [settings.STATIC_URL + 'extradmin/js/jquery.inlines.js']
+    @property
+    def media(self):
+        js = [static('extradmin/js/jquery.inlines.js')]
         if self.prepopulated_fields:
-            js.append(settings.ADMIN_MEDIA_PREFIX + 'js/urlify.js')
-            js.append(settings.STATIC_URL + 'extradmin/js/jquery.prepopulate.js')
+            js.append(static('admin/js/urlify.js'))
+            js.append(static('extradmin/js/jquery.prepopulate.js'))
         return forms.Media(js=js)
-    media = property(_media)
 
     @property
     def markup_widget(self):