+++ lemon/sitemaps/urls.py
@@ -1,4 +1,4 @@
-from django.conf.urls.defaults import patterns, url
+from django.conf.urls import patterns, url
 from lemon.sitemaps.views import sitemap_xml