+++ HISTORY.rst
@@ -1,6 +1,38 @@
 History
 =======
 
+0.6.0
+-----
+
+* Update required Django version to 1.4
+
+0.5.5
+-----
+
+* Fix sync_sitemap and sync_metatags commands
+
+0.5.4
+-----
+
+* Add SearchIndex for pages.Page
+
+0.5.3
+-----
+
+* Make inline admin class used in MetatagsSite cutomizable
+
+0.5.2
+-----
+
+* Update logos
+* Fix app titles in app dashboard widget
+
+0.5.1
+-----
+
+* Fix handling of the _db attribute on the PublicationManager in get_query_set
+* Fix robots.txt template
+
 0.5
 ---